fix(a11y): expose expand state, lengthen undo, RTL-aware login slide
Add stateDescription to collapsible meal-plan day cards so screen readers announce expanded/collapsed state, lengthen undo snackbars from 4s to 10s (SnackbarDuration.Long), and make the login registration slide direction-aware for RTL locales. Also fix the inverted visibility icon on the AI key field.

+++ composeApp/src/commonMain/kotlin/app/mymultiverse/ammo/presentation/components/JourneySnackbarHost.kt
@@ -14,7 +14,7 @@ object JourneySnackbarDefaults {
     val stickyInputBarHeight = 80.dp
 
     /** Auto-dismiss action snackbars (e.g. undo delete) unless the user taps the action. */
-    val actionSnackbarDuration = SnackbarDuration.Short
+    val actionSnackbarDuration = SnackbarDuration.Long
 }
 
 suspend fun SnackbarHostState.showJourneyActionSnackbar(
